backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
twilio
scikit-learn
numpy
pillow
python-multipart